// src/Interfaces/ICars.ts
interface ICars{
    id: number;
    brand: string;
    model: string;
    year: number;
    pricePerDay: number;
    status: string;
    img: string;
    description: string;
    transmission: string;
    kilometer: string;
    brakes: string;
    fuelType: string;
    rating: number;
    userId: number;
};

export default ICars;